Multimodal/requirements.txt
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
opencv-python==4.10.0.84
matplotlib
numpy
pandas
scipy
timm
imageio
imageio-ffmpeg
soundfile
# TensorFlow (Apple builds)
tensorflow==2.16.2
scikit-learn
scikit-image
pytest
pybind11
diffusers>=0.20.0
transformers
accelerate
mlflow
langchain
langchain-openai
openai
openpyxl
python-dotenv>=1.0.0
streamlit>=1.28.0
networkx>=3.1
pillow>=10.0.0
